Refactor Product entity: Rename CreatedBy to Manager
Updated the `Product` entity in the `DigitalRightsManagement` application by removing the `CreatedBy` property and replacing it with a new `Manager` property. The migration reflects this change by renaming the corresponding column in the `Products` table. Additionally, a new `Name` property has been added with a maximum length of 50 characters to enhance product ownership representation.