Rename "Manager" column to "UserId" in Products table
This commit updates the database schema by renaming the "Manager" column to "UserId" in the "Products" table. The migration class `ChangeNameOfProductOwnerColumn` has been modified to include the necessary changes in both the `Up` and `Down` methods. Additionally, the model snapshot has been updated to reflect this change, replacing the "Manager" property with "UserId" in the `Product` entity definition.